TESTS: Add snapshot tests - needs polishing

+++ devlxdtest/vol-mgmt-test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -ex
+set -exo
 
 poolDriver="zfs"
 instPrefix="vol-mgmt-test"
@@ -48,7 +48,11 @@ done
 # Cleanup existing volumes before the test.
 if lxc storage list --format csv | awk -F, '{ print $1}' | grep -qE "^${pool}$"; then
     for vol in $(lxc storage volume ls "${pool}" --format csv --columns n); do
-        lxc storage volume delete "${pool}" "${vol}"
+        if [[ "${vol}" == *"/"* ]]; then
+            echo "Skipping snapshot ${vol} removal"
+        else
+            lxc storage volume delete "${pool}" "${vol}"
+        fi
     done
 else
     lxc storage create "${pool}" "${poolDriver}"
@@ -57,6 +61,8 @@ fi
 # > Tests
 
 for instType in container vm; do
+    echo "==> TEST: Test devLXD volume management through a ${instType}"
+
     inst="${instPrefix}-${instType}"
 
     if [ "${instType}" = "vm" ]; then
@@ -66,6 +72,8 @@ for instType in container vm; do
     lxc launch ubuntu-minimal:24.04 "${inst}" ${ARGS}
     waitInstance "${inst}"
 
+    echo "===> TEST: Test devLXD conectivity and access"
+
     lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X GET lxd/1.0
 
     # Test unauthorized access.
@@ -74,7 +82,8 @@ for instType in container vm; do
     # Enable volume management.
     lxc config set "${inst}" security.devlxd.management.volumes=true
 
-    # > Storage pools.
+    echo "===> PASS: Test devLXD conectivity and access"
+    echo "===> TEST: Test storage pools"
 
     # List storage pools (fail - only recurison is supported).
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools || false
@@ -86,13 +95,16 @@ for instType in container vm; do
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/invalid-pool || false
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool} | jq
 
+    echo "===> PASS: Test storage pools"
+    echo "===> TEST: Test storage volumes"
+
     # Get storage volumes (fail - non-custom volume requested).
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/image || false
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/container || false
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/virtual-machine || false
 
     # Get storage volumes (ok - custom volumes requested).
-    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom | jq 'length == 0'
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom | jq -e 'length == 0'
 
     # Create storage volumes (fail - invalid or unknown type).
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X POST lxd/1.0/storage-pools/${pool}/volumes/custom -H "Content-Type: application/json" -d '{"name": "vol-type-mismatch", "type": "container"}' || false
@@ -114,7 +126,7 @@ for instType in container vm; do
     ! lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X POST lxd/1.0/storage-pools/${pool}/volumes/custom -H "Content-Type: application/json" -d '{"name": "vol-01"}' || false
 
     # Verify created storage volumes.
-    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom | jq 'length == 2'
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom | jq -e 'length == 2'
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom/vol-01 | jq | tee >(jq -e '.config.size == "10MiB" and .description == ""')
 
@@ -141,7 +153,7 @@ for instType in container vm; do
 
     # Update storage volume
     lxc exec "${inst}" -- curl --unix-socket /dev/lxd/sock \
-        -X PUT lxd/1.0/storage-pools/${pool}/volumes/custom/vol-01 \
+        -X PATCH lxd/1.0/storage-pools/${pool}/volumes/custom/vol-01 \
         -H "Content-Type: application/json" \
         -d '{"config": {"size": "25MiB"}}'
     lxc exec "${inst}" -- curl -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/storage-pools/${pool}/volumes/custom/vol-01 | jq | tee >(jq -e '.config.size == "25MiB" and .description == "Updated volume"')
@@ -152,6 +164,9 @@ for instType in container vm; do
         -H "Content-Type: application/json" \
         -H "If-Match: incorrect-etag" \
         -d '{"description": "This must fail"}' || false
+
+    echo "===> PASS: Test storage volumes"
+    echo "===> TEST: Test instance devices"
 
     # List devices.
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/instances/${inst}
@@ -186,6 +201,38 @@ EOF
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X PUT lxd/1.0/instances/${inst} -H "Content-Type: application/json" -d "${detachReq}"
     lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET lxd/1.0/instances/${inst} | jq '.devices' | jq 'length == 0'
 
+    echo "===> PASS: Test instance devices"
+    echo "===> TEST: Test storage volume snapshots"
+
+    # Create storage volume snapshots.
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots?recursion=1" | jq -e 'length == 0'
+
+    opID=$(lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X POST "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots" \
+        -H "Content-Type: application/json" \
+        -d '{"name": "snap-01"}' | jq -er .id)
+    lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X GET "lxd/1.0/operations/${opID}/wait?timeout=10"
+
+    opID=$(lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X POST "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots" \
+        -H "Content-Type: application/json" \
+        -d '{"name": "snap-02"}' | jq -er .id)
+    lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X GET "lxd/1.0/operations/${opID}/wait?timeout=10"
+
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots?recursion=1"
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots?recursion=1" | jq -e 'length == 2'
+
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots/snap-01" | jq -e '.name == "snap-01" and .content_type == "filesystem"'
+
+    # Delete storage volume snapshot.
+    opID=$(lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X DELETE "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots/snap-01" | jq -er .id)
+    lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X GET "lxd/1.0/operations/${opID}/wait?timeout=10"
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots?recursion=1" | jq -e 'length == 1'
+    opID=$(lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X DELETE "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots/snap-02" | jq -er .id)
+    lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X GET "lxd/1.0/operations/${opID}/wait?timeout=10"
+    lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots?recursion=1" | jq -e 'length == 0'
+    ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X GET "lxd/1.0/storage-pools/${pool}/volumes/custom/vol-02/snapshots/snap-02" || false
+
+    echo "===> PASS: Test storage volume snapshots"
+
     # Delete storage volumes.
     ! lxc exec "${inst}" -- curl -f -s --unix-socket /dev/lxd/sock -X DELETE lxd/1.0/storage-pools/${pool}/volumes/custom/non-existing-volume || false
     lxc exec "${inst}" -- curl -f --unix-socket /dev/lxd/sock -X DELETE lxd/1.0/storage-pools/${pool}/volumes/custom/vol-01
@@ -196,6 +243,8 @@ EOF
 
     # Cleanup.
     lxc delete "${inst}" --force
+
+    echo "==> PASS: Test devLXD volume management through a ${instType}"
 done
 
 # Cleanup.
